wip: archivedBid
