+++ webapp/src/sagas.js
@@ -3,12 +3,12 @@ import { createTranslationSaga } from '@dapps/modules/translation/sagas'
 import { addressSaga } from 'modules/address/sagas'
 import { analyticsSaga } from '@dapps/modules/analytics/sagas'
 import { transactionSaga } from '@dapps/modules/transaction/sagas'
+import { locationSaga } from '@dapps/modules/location/sagas'
 
 import { assetSaga } from 'modules/asset/sagas'
 import { auctionSaga } from 'modules/auction/sagas'
 import { authorizationSaga } from 'modules/authorization/sagas'
 import { districtsSaga } from 'modules/districts/sagas'
-import { locationSaga } from '@dapps/modules/location/sagas'
 import { parcelsSaga } from 'modules/parcels/sagas'
 import { publicationSaga } from 'modules/publication/sagas'
 import { walletSaga } from 'modules/wallet/sagas'
